requirements.txt
argoscuolanext==2.0.1
colorama==0.4.6
playsound==1.3.0
python-dotenv==1.0.1
readchar==4.0.5
Requests==2.31.0
